Bugfix/SMAR-11932

Approved-by: Sandip.Kumar <sandip.kumar@tetrapak.com>

+++ tetrapak-publicweb/core/src/main/java/com/tetrapak/publicweb/core/services/impl/PseudoCategoryServiceImpl.java
@@ -92,7 +92,7 @@ public class PseudoCategoryServiceImpl implements PseudoCategoryService {
         cfBean.setPseudoCategoryKey(valueMap.get("pseudoCateoryKey", StringUtils.EMPTY));
         cfBean.setPseudoCategoryValue(valueMap.get("pseudoCateoryValue", StringUtils.EMPTY));
         if (valueMap.containsKey("pseudoCateoryOrder")) {
-            cfBean.setPseudoCategoryOrder(((Double) valueMap.get("pseudoCateoryOrder")).intValue());
+            cfBean.setPseudoCategoryOrder(((Long) valueMap.get("pseudoCateoryOrder")).intValue());
         }
         return cfBean;
     }
